add date to next birthdays info
