+++ src/models/birthday.py
@@ -1,0 +1,10 @@
+from src.models.field import Field
+
+
+class Birthday(Field):
+    def __init__(self, value):
+        try:
+            self.value = datetime.strptime(value, "%d.%m.%Y").date()
+        except ValueError:
+            raise ValueError("Invalid date format. Use DD.MM.YYYY")
+        super().__init__(self.value)
